portals/xfo/portals-xfo-sameorigin.sub.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
  assert_implements("HTMLPortalElement" in self);
  var portal = document.createElement('portal');
  portal.src = "http://{{domains[www]}}:{{ports[http][0]}}/portals/xfo/resources/xfo-sameorigin.asis";
  portal.onmessage = t.unreached_func("should not have received a message");
  document.body.appendChild(portal);
  t.add_cleanup(() => portal.remove());
  t.step_timeout(() => t.done(), 2000);
}, "`XFO: SAMEORIGIN` blocks cross-origin portals.");
</script>
</body>
